Refactor tone generation to allow setting both frequency and duration. The JSON API now accepts frequency and duration as separate optional values, defaulting to last used if either are not provided. This means it's now enough to send a empty tone payload `{"tone":{}}` and the tone will play.

+++ include/avr/ToneHandler.h
@@ -14,7 +14,7 @@ private:
 public:
     void begin();
 
-    static void play(unsigned int frequency, unsigned long duration);
+    static void play(unsigned int frequency, unsigned int duration);
 };
 
 #endif // ARDUINO_ARCH_AVR
